+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## Unreleased
 
-Bugs:
+Security:
 * multi-line secrets are now properly masked in logs [GH-208](https://github.com/hashicorp/vault-action/pull/208)
 
 ## 2.1.2 (January 21st, 2021)