+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## Unreleased
 
+Features:
+* JWT auth method is now supported [GH-188](https://github.com/hashicorp/vault-action/pull/188)
+
 ## 2.1.2 (January 21st, 2021)
 
 Bugs: