+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 Security:
 * multi-line secrets are now properly masked in logs [GH-208](https://github.com/hashicorp/vault-action/pull/208)
 
+Features:
+* JWT auth method is now supported [GH-188](https://github.com/hashicorp/vault-action/pull/188)
+
 ## 2.1.2 (January 21st, 2021)
 
 Bugs: